# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample doc for mod-checkup.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.7.3"
  title: "Mod-Shutup API Docs"
host: "localhost:5000"
basePath: "/api"
...
schemes:
- "http"
- "https"
paths:
...
  /subjects:
    post:
      tags:
      - "subjects"
      summary: "Add a new subject"
      operationId: "addSubject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "Subject that you want to create"
        required: true
        schema:
          $ref: "#/definitions/Subject"
      responses:
        "201":
          description: "Created subject will be returned"
        "409":
          description: "Something went wrong"
    get:
      tags:
      - "subjects"
      summary: "Get all subjects"
      operationId: "getAllSubjects"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "All of the subjects will be returned"
        "404":
          description: "Not Found"
  /subjects/{subjectAbbr}:
    get:
      tags:
      - "subjects"
      summary: "Get subject info by subject"
      operationId: "getSubjectInfo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "subjectAbbr"
        in: "path"
        type: "string"
        description: "Subject Abbreviation (Ex. OOP122)"
        required: true
      responses:
        "200":
          description: "Specified subject's infos will be returned"
        "404":
          description: "Not Found"
    put:
      tags:
      - "subjects"
      summary: "Update subject info"
      operationId: "updateSubject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "subjectAbbr"
        in: "path"
        type: "string"
        description: "Subject Abbreviation (Ex. OOP122)"
        required: true
      - name: "body"
        in: "body"
        description: "Subject that you want to update"
        required: true
        schema:
          $ref: "#/definitions/Subject"
      responses:
        "200":
          description: "Updated subject"
        "404":
          description: "Not Found"
...
  /users:
    get:
      tags:
      - "users"
      summary: "Get all users"
      operationId: "getAllUsers"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "All users will be returned"
        "404":
          description: "Not Found"
  /users/{userId}:
    get:
      tags:
      - "users"
      summary: "Get specified user info"
      operationId: "findUserById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        type: "string"
        required: true
      responses:
        "200":
          description: "Specified user's info will be returned"
        "404":
          description: "Not Found"
  /users/current:
    get:
      tags:
      - "users"
      summary: "Get current user info"
      operationId: "getCurrentUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Current user's info will be returned"
        "401":
          description: "Unauthorized access"
...
securityDefinitions:
  google_oauth2:
    type: "oauth2"
    authorizationUrl: "http://localhost:5000/api/auth/google"
    flow: "implicit"
    scopes:
      read: "Grants read access"
      write: "Grants write access"
      admin: "Grants read and write access to administrative information"
definitions:
...
  Subject:
    type: "object"
    properties:
      subject_abbr:
        type: "string"
        description: "Subject Abbreviation (FDM012)"
      subject_name:
        type: "string"
        description: "Subject full name (Fire Department Managing 1)"
      average_rating:
        type: "number"
        description: "Average rating of subject"
        minimum: 0
        maximum: 5
    required:
      - subject_abbr
      - subject_name
...